pyproject.toml
[project]
name = "ipps"
version = "0.1.0"
description = ""
authors = [
    {name = "macaulaypraise",email = "Macaulaypraise37@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.133.1,<0.134.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "sqlalchemy (>=2.0.47,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "redis[hiredis] (>=7.2.1,<8.0.0)",
    "aiokafka (>=0.13.0,<0.14.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "prometheus-client (>=0.24.1,<0.25.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.2,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-mock (>=3.15.1,<4.0.0)",
    "locust (>=2.43.3,<3.0.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "pytest-dotenv (>=0.5.2,<0.6.0)"
]

[tool.poetry]
package-mode = false

[tool.ruff]
# Target your specific pinned version
target-version = "py312"
line-length = 88

[tool.ruff.lint]
# Enable Pyflakes (F), pycodestyle (E), isort (I), and pyupgrade (UP)
select = ["E", "F", "I", "UP"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]

[tool.mypy]
python_version = "3.12"
plugins = [
  "pydantic.mypy"
]
# Tell Mypy to completely ignore the tests directory
exclude = ["^tests/"]

# FastAPI relies heavily on Pydantic and strict type hints.
# This enforces industry-standard type safety.
strict = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_return_any = true
warn_unused_ignores = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

# Suppress errors for third-party packages missing type stubs
[[tool.mypy.overrides]]
module = "aiokafka.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.core.redis_client"
disable_error_code = ["misc"]

[[tool.mypy.overrides]]
module = "tests.load.locustfile"
disable_error_code = ["no-untyped-call"]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# Uncomment and modify this block if you install a third-party package
# that does not have built-in type hints to prevent Mypy from crashing.
# [[tool.mypy.overrides]]
# module = "some_untyped_package.*"
# ignore_missing_imports = true

[tool.pytest.ini_options]
# Define where your tests will live
testpaths = ["tests"]
env_files = [".env.test"]

# Automatically detect and run async tests
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

# Default command line options when you run `pytest`
# -v: verbose output
# --cov: generate a coverage report for your main code folder (assuming it's named 'app')
# --cov-report: show exactly which lines of code are missing tests in the terminal
addopts = "-v --cov=app --cov-report=term-missing"
